@@ -1,4 +1,5 @@
 """Browser automation extension for Goose."""
-from .server import main
+from browser_automation.server import main
 
+__version__ = "0.1.0"
 __all__ = ['main']
